Create and Structure index, style and script
Madee the three core files. Populated the
main DOM in the index to start the game and
made the created the startGame module. Also
structured a bit of the displayController
module and the stats module. displayController
is dependant on the startGame module.

start game. Also created a pubsub which I
want to use to communicate changes to the
sta